Figured it out!
I was looking for things that would change instead of static stuff.
Anyway, the problem was that we stored a static list of metrics in the
SnackBar undo callback. If the user added items before undoing, we would
set the wrong items and incorrectly notify the adapter. The solution is
use the latest items and insert the deleted item.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>