Fix fab disappearing when adding team in tablet mode even if there aren't enough
When adding a team in tablet mode, the config changes and squishes itself
to accommodate the keyboard. That's good, but then if a new team is added,
the list is long enough to require a scroll which hides the FAB. However,
once the keyboard is gone again, the list isn't long enough to scroll
and the FAB will be stuck in its hidden state forever.

Adding the show thing to the menu options callback means selecting a team
will show the FAB, but that's acceptable behavior.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>