Move team merging logic to server
This commit also fixes a bug where only one merge could occur per session. The app would have to be
restarted to process other merges. In addition, because we are merging teams on the server, we can
now merge teams across accounts. Previously, if a team had 2 or more owners, it would never be
merged.

This change has the following benefits:
- Multiple devices don't step over each other trying to merge the teams
- On-device performance is improved since we no longer have to iterate through all teams every time
  something changes
- Possible race conditions are eliminated

In addition, team merging no longer cares if the teams are the same or not. This change was done to
minimize confusion. Furthermore, since Auto Scout will pre-fetch all teams in advance to the event,
there should be close to no duplicate teams. A side effect of this change is that pending media
uploads or share tokens will be lost. This is acceptable given the aforementioned reasoning and low
frequency of merges.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>