Revamp the entire coroutines error logging situation (closes #249)
Currently, we have to capture a stack trace regardless of whether or not an error will be thrown. This is bad for a number of reasons: doesn't work on pre-L devices, negative performance impact, hacky, etc. In the new method, we get potentially less accurate stack traces, but it's a much better way of doing things. The `*.await()` caller is catches exceptions and rethrows a bridging exception if it doesn't want to deal with the caught exception. This is good because it requires thinking about exceptions and manages to capture the try/catch block in which the exception was thrown.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>